Added binary flag in init method for Homegear

+++ addons/binding/org.openhab.binding.homematic/src/main/java/org/openhab/binding/homematic/internal/communicator/client/RpcClient.java
@@ -80,7 +80,7 @@ public abstract class RpcClient<T> {
         request.addArg(getRpcCallbackUrl());
         request.addArg(clientId);
         if (config.getGatewayInfo().isHomegear()) {
-            request.addArg(new Integer(0x20));
+            request.addArg(new Integer(0x22));
         }
         sendMessage(config.getRpcPort(hmInterface), request);
     }
